PR reviews comments and Hibernate 6.3+ compatible issue resolution
